Move CutRepresentation class to separate file
